services/factory-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
python-multipart==0.0.6
google-cloud-pubsub==2.18.4
google-cloud-secret-manager==2.16.4
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
alembic==1.13.1
cryptography==42.0.0
pytest==7.4.3
pytest-asyncio==0.23.2
ruff==0.1.9
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
